pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "regfmt"
version = "0.9.0"
authors = [
  { name="Charles Choi", email="kickingvegas@gmail.com"}
]

description = "A command line utility to generate SVG diagrams for control register-style data formats."
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]

dependencies = [
    "attrs>=22.1.0",
    "jsonschema>=4.16.0",
    "Pillow>=9.2.0",
    "pyrsistent>=0.18.1",
    "PyYAML>=6.0",
    "tinycss2>=1.1.1",
    "webencodings>=0.5.1"
]

[project.urls]
"Homepage" = "https://github.com/kickingvegas/regfmt"
"Issue Tracker" = "https://github.com/kickingvegas/regfmt/issues"

[project.scripts]
regfmt = "regfmt:main"
